heartbeat-monitor/requirements.txt
requests
paho-mqtt
python-dotenv
google-cloud-logging